Phyton/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
firebase-admin==6.2.0
Gunicorn==21.2.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.41.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
